+++ packages/components/src/components/header/header.component.ts
@@ -20,18 +20,21 @@ import { watch } from '../../internal/watch.js';
  * @status stable
  * @since 1.10.0
  *
- * @slot - The label for the header.
- * @slot logo - The logo that should be displayed. Will fall back to the SICK logo if not provided.
- * @slot meta-navigation - The meta-navigation is used to add various application toolbar icons.
+ * @slot label - The label for the header
+ * @slot logo - The logo that should be displayed. Will fall back to the SICK logo if not provided
+ * @slot meta-navigation - The meta-navigation is used to add various application toolbar icons
  *                     Best used with `<syn-icon-button />` and `<syn-drop-down />`
  * @slot navigation - This slot can be used to add an optional horizontal navigation
- * @slot show-burger-menu - An icon to use in lieu of the default show burger menu icon
- * @slot hide-burger-menu - An icon to use in lieu of the default hide burger menu icon
+ * @slot open-burger-menu-icon - An icon to use in lieu of the default burger-menu=open state.
+ *                      The default close icon is a 'x'.
+ * @slot closed-burger-menu-icon - An icon to use in lieu of the default burger-menu=closed state.
+ *                      The default open icon is a burger menu.
  *
- * @event syn-burger-menu-show - Emitted when the burger menu button is toggled to visible
- * @event syn-burger-menu-hide - Emitted when the burger menu button is toggled to not visible
+ * @event syn-burger-menu-closed - Emitted when the burger menu is toggled to closed
+ * @event syn-burger-menu-hidden - Emitted when the burger menu is toggled to hidden
+ * @event syn-burger-menu-open - Emitted when the burger menu is toggled to open
  *
- * @csspart base - The component's base wrapper.
+ * @csspart base - The component's base wrapper
  * @csspart content - The wrapper most content items reside
  * @csspart logo - The wrapper the application logo resides in
  * @csspart label - The element wrapping the application name
@@ -59,46 +62,60 @@ export default class SynHeader extends SynergyElement {
   private mutationObserver: MutationObserver;
 
   /**
-   * The headers label. If you need to display HTML, use the `default` slot instead.
+   * The headers label. If you need to display HTML, use the `label` slot instead.
    */
   @property() label = '';
 
   /**
-   * Adds a button to toggle the burger menu's visibility.
-   * The button is added automatically, if the component finds a syn-side-nav in non-rail mode.
+   * Defines the current visibility and icon of the burger-menu icon.
+   * The menu button is added automatically if the component finds a syn-side-nav in non-rail mode.
+   * The following values can be used:
+   * - hidden: The burger menu is not visible
+   * - open: The burger menu is visible and shows the close icon
+   * - closed: The burger menu is visible and shows the open icon
    */
-  @property({ attribute: 'show-burger-menu', reflect: true, type: Boolean }) showBurgerMenu = false;
-
-  /**
-   * Determines whether or not the burger menu is currently visible.
-   */
-  @property({ attribute: 'burger-menu-visible', reflect: true, type: Boolean }) burgerMenuVisible = false;
+  @property({ attribute: 'burger-menu', reflect: true }) burgerMenu: 'hidden' | 'open' | 'closed' = 'hidden';
 
   /**
    * The side nav
    */
   @state() private sideNav: SynSideNav | null;
 
+  private toggleBurgerMenu() {
+    switch (this.burgerMenu) {
+    case 'closed': this.burgerMenu = 'open'; break;
+    case 'open': this.burgerMenu = 'closed'; break;
+    default: break;
+    }
+  }
+
   private handleBurgerMenuToggle() {
     // If there is a side-nav in non-rail mode, toggle the open state!
     if (this.sideNav && !this.sideNav.rail) {
       this.sideNav.open = !this.sideNav.open;
     }
-    this.burgerMenuVisible = !this.burgerMenuVisible;
+    this.toggleBurgerMenu();
   }
 
+  /**
+   * Automatically update the burger menu icon based
+   * on the state of the side-nav, if one is connected.
+   */
   private updateBurgerMenuBasedOnSideNav() {
-    // The side-nav has a higher priority than when the burger menu toggle and visibility
-    // is set via property
     if (this.sideNav) {
-      this.showBurgerMenu = !this.sideNav.rail;
-      this.burgerMenuVisible = !this.sideNav.rail ? this.sideNav.open : false;
+      // Hide the burger menu icon if the side-nav is in rail mode
+      if (this.sideNav.rail) {
+        this.burgerMenu = 'hidden';
+      } else {
+        this.burgerMenu = this.sideNav.open ? 'open' : 'closed';
+      }
     }
   }
 
-  @watch('burgerMenuVisible', { waitUntilFirstUpdate: true })
-  handleBurgerMenuVisible() {
-    this.emit(this.burgerMenuVisible ? 'syn-burger-menu-show' : 'syn-burger-menu-hide');
+  @watch('burgerMenu', { waitUntilFirstUpdate: true })
+  handleBurgerMenu() {
+    const myEvt: keyof GlobalEventHandlersEventMap = `syn-burger-menu-${this.burgerMenu}`;
+    this.emit(myEvt);
   }
 
   connectedCallback() {
@@ -110,7 +127,7 @@ export default class SynHeader extends SynergyElement {
   firstUpdated() {
     // Search for a side-nav and use the first found in case of the connectSideNavigation method
     // is not used by the user.
-    const sideNav = document.querySelector('syn-side-nav')!;
+    const sideNav = document.querySelector('syn-side-nav');
     this.connectSideNavigation(sideNav);
   }
 
@@ -130,7 +147,7 @@ export default class SynHeader extends SynergyElement {
    *
    * @param sideNav The side navigation to connect to the header
    */
-  connectSideNavigation(sideNav: SynSideNav) {
+  connectSideNavigation(sideNav: SynSideNav | null) {
     this.mutationObserver.disconnect();
 
     this.sideNav = sideNav || document.querySelector('syn-side-nav');
@@ -144,12 +161,12 @@ export default class SynHeader extends SynergyElement {
 
   render() {
     const hasNavigation = this.hasSlotController.test('navigation');
-
+    const showBurgerMenu = this.burgerMenu !== 'hidden';
     return html`
       <header
         class=${classMap({
           header: true,
-          'header--has-burger-menu': this.showBurgerMenu,
+          'header--has-burger-menu': showBurgerMenu,
           'header--has-navigation': hasNavigation,
         })}
         part="base"
@@ -157,23 +174,23 @@ export default class SynHeader extends SynergyElement {
         <!-- .header__content -->
         <div part="content" class="header__content">
 
-          ${this.showBurgerMenu
+          ${showBurgerMenu
               ? html`
                   <button
-                    part="burger-menu-toggle-button"
+                    aria-label=${this.localize.term(this.burgerMenu === 'closed' ? 'openMenu' : 'closeMenu')}
                     class="header__burger-menu-toggle"
-                    type="button"
                     @click=${this.handleBurgerMenuToggle}
-                    aria-label=${this.localize.term(this.burgerMenuVisible ? 'closeMenu' : 'openMenu')}
+                    part="burger-menu-toggle-button"
+                    type="button"
                   >
-                    ${this.burgerMenuVisible
+                    ${this.burgerMenu === 'open'
                       ? html`
-                          <slot name="show-burger-menu">
+                          <slot name="open-burger-menu-icon">
                             <syn-icon name="x-lg" library="system"></syn-icon>
                           </slot>
                         `
                       : html`
-                          <slot name="hide-burger-menu">
+                          <slot name="closed-burger-menu-icon">
                             <syn-icon name="menu" library="system"></syn-icon>
                           </slot>
                         `}
@@ -188,7 +205,7 @@ export default class SynHeader extends SynergyElement {
           </div>
 
           <div part="label" class="header__label">
-            <slot>
+            <slot name="label">
               ${this.label}
             </slot>
           </div>