+++ packages/components/src/events/syn-burger-menu-open.ts
@@ -1,0 +1,7 @@
+export type SynBurgerMenuOpenEvent = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'syn-burger-menu-open': SynBurgerMenuOpenEvent;
+  }
+}