+++ packages/vue/src/components/SynVueBreadcrumbItem.vue
@@ -31,13 +31,7 @@ import type { SynBreadcrumbItem } from '@synergy-design-system/components';
 // DOM Reference to the element
 const nativeElement = ref<SynBreadcrumbItem>();
 
-// Map methods
-const callHrefChanged = (...args: Parameters<SynBreadcrumbItem['hrefChanged']>) => nativeElement.value?.hrefChanged(...args);
-const callHandleSlotChange = (...args: Parameters<SynBreadcrumbItem['handleSlotChange']>) => nativeElement.value?.handleSlotChange(...args);
-
 defineExpose({
-  callHrefChanged,
-  callHandleSlotChange,
   nativeElement,
 });