+++ packages/vue/src/components/SynVueButtonGroup.vue
@@ -23,10 +23,7 @@ import type { SynButtonGroup } from '@synergy-design-system/components';
 // DOM Reference to the element
 const nativeElement = ref<SynButtonGroup>();
 
-// Map methods
-
 defineExpose({
-
   nativeElement,
 });