+++ packages/components/CHANGELOG.md
@@ -1,3 +1,27 @@
+# [@synergy-design-system/components-v2.0.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/1.27.0...components/2.0.0) (2024-06-05)
+
+
+### Bug Fixes
+
+* 🤔 v2 breaking changes ([#448](https://github.com/synergy-design-system/synergy-design-system/issues/448)) ([b251fa7](https://github.com/synergy-design-system/synergy-design-system/commit/b251fa72ec6668d005eed23c561e901dc050ab83))
+
+
+### BREAKING CHANGES
+
+* syn-header - Deprecate default slot in favor of label slot
+* syn-header - fix attribute naming
+* syn-prio-nav - Deprecated prop "priority-menu-label"
+* syn-side-nav - Adjust default width of open sidenav
+* syn-card - Rename nested prop to sharp
+* Wrappers - Remove deprecated methods in wrappers
+
+# [@synergy-design-system/components-v1.27.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/1.26.3...components/1.27.0) (2024-06-05)
+
+
+### Features
+
+* ✨ icon-only button ([#498](https://github.com/synergy-design-system/synergy-design-system/issues/498)) ([d223e2c](https://github.com/synergy-design-system/synergy-design-system/commit/d223e2cd649fce2bb9109798613d4b7c60ce2117))
+
 # [@synergy-design-system/components-v1.26.3](https://github.com/synergy-design-system/synergy-design-system/compare/components/1.26.2...components/1.26.3) (2024-06-05)